Implemented splash screen
