+++ pubspec.yaml
@@ -45,6 +45,7 @@ flutter:
   # - load_json/person.json
     - images/atri.jpg
     - images/ritam.jpeg
+    - images/logo.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.io/assets-and-images/#resolution-aware.
